# Conflicts:
#	Depo/Depo/Presentation/UserStories/Settings/ContactSync/MainScreen/ContactSyncViewController.swift

+++ Depo/Depo/Presentation/UserStories/Settings/ContactSync/MainScreen/Views/ContactSyncProgressView.swift
@@ -31,7 +31,7 @@ final class ContactSyncProgressView: UIView, NibInit {
         }
     }
     
-    func update(progress: Float) {
+    func update(progress: Int) {
         loader.progress = CGFloat(progress)
     }
 }
